+++ src/modules/index.mjs
@@ -7,9 +7,9 @@ import ImportProcessor from "./processor";
 
 import LogModule from "@prisma-cms/log-module";
 import UserModule from "@prisma-cms/user-module";
-import ResourceModule from "@prisma-cms/resource-module";
-import SocietyModule from "@prisma-cms/society-module";
-import CooperationModule from "@prisma-cms/cooperation-module";
+// import ResourceModule from "@prisma-cms/resource-module";
+// import SocietyModule from "@prisma-cms/society-module";
+// import CooperationModule from "@prisma-cms/cooperation-module";
 
 import PrismaModule from "@prisma-cms/prisma-module";
 
@@ -39,10 +39,10 @@ class Module extends PrismaModule {
 
     this.mergeModules([
       LogModule,
-      SocietyModule,
       UserModule,
-      ResourceModule,
-      CooperationModule,
+      // SocietyModule,
+      // ResourceModule,
+      // CooperationModule,
     ]);
 
   }